ADD double session du 05/04/2022

+++ concat.sh
@@ -16,6 +16,11 @@ DESCRIPTION
 !
 }
 
+if [ "$#" -ne 1 ]; then
+    usage
+    exit 0;
+fi
+
 case $TARGET in
     speaker)
         rm ${TARGET}.txt || true
